refactor: move window to own file
